+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - Prompt the user for a password if the database password is not set
 - Generate only lowercase values in RandomizeText and RandomizeEmail converters
+- YAML validation: allow number values when a string is expected
 
 ## [1.0.0-beta16] - 2019-12-17
 [1.0.0-beta16]: https://git.smile.fr/dirtech/gdpr-dump/compare/1.0.0-beta15...1.0.0-beta16